feat: refine user role assignment logic in UserService and enhance user manipulation checks in UserCRUDModelViewSet

+++ backend/apps/authentication/views.py
@@ -69,10 +69,7 @@ class UserCRUDModelViewSet(ModelViewSet):
     def get_queryset(self):
         return (
             User.objects
-            .prefetch_related('groups')
-            .filter(is_superuser=False)
-            .exclude(groups__name='admin')
-            .distinct()
+            .prefetch_related('groups').all()
         )
 
     filter_backends = [DjangoFilterBackend]
@@ -135,6 +132,10 @@ class UserCRUDModelViewSet(ModelViewSet):
     responses={200: UserReadSerializer}
     )
     def update(self, request, *args, **kwargs):
+        target_user = self.get_object()
+        target_groups = target_user.groups.values_list('name', flat=True).first()
+        if target_user.is_superuser or target_groups == 'admin':
+            return Response("You can not manipulate this user", status=status.HTTP_400_BAD_REQUEST)
         return super().update(request, *args, **kwargs)
     
     @extend_schema(
@@ -144,6 +145,10 @@ class UserCRUDModelViewSet(ModelViewSet):
     responses={200: UserReadSerializer}
     )
     def partial_update(self, request, *args, **kwargs):
+        target_user = self.get_object()
+        target_groups = target_user.groups.values_list('name', flat=True).first()
+        if target_user.is_superuser or target_groups == 'admin':
+            return Response("You can not manipulate this user", status=status.HTTP_400_BAD_REQUEST)
         return super().partial_update(request, *args, **kwargs)
 
 
@@ -157,6 +162,9 @@ class UserCRUDModelViewSet(ModelViewSet):
     )
     def destroy(self, request, *args, **kwargs):
         target_user = self.get_object()
+        target_groups = target_user.groups.values_list('name', flat=True).first()
+        if target_user.is_superuser or target_groups == 'admin':
+            return Response("You can not manipulate this user", status=status.HTTP_400_BAD_REQUEST)
         try:
             UserService.delete_user(request.user, target_user)
         except ValueError as e:
